Conveyors
Needs animations and Zach needs to cause the player to move with the conveyor.

Conveyor prefab found in prefab - matthew - resources

Place a conveyor in the scene. It will move by default unless there is an ElectActivate gameobject with the conveyor as its "go" and doConvey = true. Then it will only convey if the electactivator is electrified.